SGEA_Project/requirements.txt
Django
reportlab
Pillow
djangorestframework